Backend/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
mysql-connector==2.2.9
packaging==24.1
pluggy==1.5.0
psycopg==3.2.1
psycopg2==2.9.9
PyJWT==2.8.0
pytest==8.3.2
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
